Bot/DiscordBot/requirements.txt
discord.py[voice]>=2.0.0
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
dashscope>=1.14.0
numpy>=1.24.0
faster-whisper>=0.9.0
edge-tts>=6.1.0
groq>=0.4.0